made json file for self assigning roles. Made command subject.
Subject command is going to be an admin only privilege. Where it will create a subject channel and subject role where only those with the role can view the channel. Currently I've made it that you can use the command to append a role name and emoji for the bot to use for roleclaim (reaction emoji self assigning role).